# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CONVIVIUM API",
    "version": "0.1.0",
    "description": "Modo edge"
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/cash/sessions": {
      "post": {
        "tags": [
          "caja"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "registerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "openingFloat": {
                    "type": "integer",
                    "description": "Monto en centavos MXN"
                  }
                },
                "required": [
                  "registerId",
                  "openingFloat"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/cash/sessions/current/movements": {
      "post": {
        "tags": [
          "caja"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "retiro",
                      "entrada"
                    ]
                  },
                  "amount": {
                    "type": "integer",
                    "description": "Monto en centavos MXN"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "type",
                  "amount",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/cash/checks/{id}/pay": {
      "post": {
        "tags": [
          "caja"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "payments": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string",
                          "enum": [
                            "efectivo_mxn",
                            "efectivo_usd",
                            "tarjeta",
                            "transferencia",
                            "vales"
                          ]
                        },
                        "amount": {
                          "type": "integer",
                          "description": "Monto en centavos MXN"
                        },
                        "exchangeRate": {
                          "type": "integer"
                        },
                        "reference": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "amount"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1
                  },
                  "tip": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "integer",
                        "description": "Monto en centavos MXN"
                      },
                      "method": {
                        "type": "string",
                        "enum": [
                          "efectivo_mxn",
                          "efectivo_usd",
                          "tarjeta",
                          "transferencia",
                          "vales"
                        ]
                      }
                    },
                    "required": [
                      "amount",
                      "method"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "payments"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "paid": {
                      "type": "integer",
                      "description": "Monto en centavos MXN"
                    },
                    "change": {
                      "type": "integer",
                      "description": "Monto en centavos MXN"
                    },
                    "checkStatus": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "paid",
                    "change",
                    "checkStatus"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/cash/checks/{id}/split": {
      "post": {
        "tags": [
          "caja"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "mode": {
                        "type": "string",
                        "enum": [
                          "iguales"
                        ]
                      },
                      "parts": {
                        "type": "integer",
                        "minimum": 2
                      }
                    },
                    "required": [
                      "mode",
                      "parts"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "mode": {
                        "type": "string",
                        "enum": [
                          "por_comensal"
                        ]
                      }
                    },
                    "required": [
                      "mode"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "mode": {
                        "type": "string",
                        "enum": [
                          "por_producto"
                        ]
                      },
                      "groups": {
                        "type": "array",
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          }
                        }
                      }
                    },
                    "required": [
                      "mode",
                      "groups"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/cash/sessions/current/counts": {
      "post": {
        "tags": [
          "caja"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "X",
                      "Z"
                    ]
                  },
                  "counted": {
                    "type": "object",
                    "required": [
                      "efectivo_mxn",
                      "efectivo_usd",
                      "tarjeta",
                      "transferencia",
                      "vales"
                    ],
                    "properties": {
                      "efectivo_mxn": {
                        "type": "integer",
                        "description": "Monto en centavos MXN"
                      },
                      "efectivo_usd": {
                        "type": "integer",
                        "description": "Monto en centavos MXN"
                      },
                      "tarjeta": {
                        "type": "integer",
                        "description": "Monto en centavos MXN"
                      },
                      "transferencia": {
                        "type": "integer",
                        "description": "Monto en centavos MXN"
                      },
                      "vales": {
                        "type": "integer",
                        "description": "Monto en centavos MXN"
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": [
                  "kind",
                  "counted"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/cash/checks/{id}/reopen": {
      "post": {
        "tags": [
          "caja"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ]
}